requirements.txt
playwright
pandas
beautifulsoup4
streamlit
supabase
python-dotenv
